refactoring and actions panel - so many changes I couldn't desribe
