+++ src/App.tsx
@@ -1,15 +1,16 @@
 import { FC } from 'react';
-
+import { useKeyboardEvent } from '@/hooks';
+import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 import { ThemeProvider } from '@/providers/ThemeProvider';
 import { OutputProvider } from '@/providers/OutputProvider';
 import { CollectionProvider } from '@/providers/CollectionProvider';
 import { SearchQueryProvider } from '@/providers/SearchQueryProvider';
+import { Header } from '@/components/header/Header';
+import { ActionsPanelRoot } from '@/components/actions-panel/ActionsPanelRoot';
+import { CollectionExplorer } from '@/components/collection-explorer/CollectionExplorer';
+import { RenderActions } from './components/actions-panel/RenderActions';
+import { ThemeActionButton } from './components/actions-panel/ThemeActionButton';
 
-import { Header } from '@/components/header';
-import { Sidebar } from '@/components/sidebar';
-import { CollectionExplorer } from '@/components/collection-explorer';
-import { useKeyboardEvent } from '@/hooks';
-import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 const appWindow = getCurrentWebviewWindow();
 
 export const App: FC = () => {
@@ -29,7 +30,11 @@ export const App: FC = () => {
               <div className="grid grid-cols-[3.5rem_1fr_3.5rem] gap-3 overflow-hidden px-2">
                 <div />
                 <CollectionExplorer />
-                <Sidebar />
+                <ActionsPanelRoot />
+
+                <RenderActions>
+                  <ThemeActionButton />
+                </RenderActions>
               </div>
             </div>
           </SearchQueryProvider>
