+++ src/assets/fonts/index.ts
@@ -1,0 +1,6 @@
+export { default as kiwiMaruFont } from './kiwi-maru.ttf';
+export { default as notoSansJpFont } from './noto-sans-jp.ttf';
+export { default as notoSerifJpFont } from './noto-serif-jp.ttf';
+export { default as segoeUiFont } from './segoe-ui.ttf';
+export { default as zenAntiqueSoftFont } from './zen-antique-soft.ttf';
+export { default as zenKurenaidoFont } from './zen-kurenaido.ttf';
