+++ src/components/actions-panel/ActionsPanelRoot.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+export const ActionsPanelRoot: FC = () => (
+  <aside id="actions-panel-root" className="flex flex-col-reverse gap-2 py-2" />
+);
