--- src/components/collection-explorer/index.ts
@@ -1,1 +1,0 @@
-export { CollectionExplorer } from './CollectionExplorer';
